2 code deployyed

+++ src/components/ProductList.tsx
@@ -1,10 +1,11 @@
 import React, { useEffect, useState } from 'react';
 import { useNavigate } from 'react-router-dom';
-import { Container, Typography, Button, IconButton } from '@mui/material';
+import { Container, Typography, Button, IconButton, Table, TableBody, TableCell, TableContainer, TableHead, TableRow, Paper,Tooltip } from '@mui/material';
 import { Edit, Delete } from '@mui/icons-material';
 import { collection, getDocs, deleteDoc, doc } from 'firebase/firestore';
 import { db } from '../firebaseConfig';
 import './ProductList.scss';
+import { FaEye } from 'react-icons/fa';
 
 interface Product {
   id: string;
@@ -12,7 +13,6 @@ interface Product {
   productCategory: string;
   productSpecCategory: string;
   productImage: string;
-  added_at: number;
 }
 
 interface Category {
@@ -34,12 +34,13 @@ const ProductList: React.FC = () => {
   const [loading, setLoading] = useState<boolean>(true);
   const [error, setError] = useState<string | null>(null);
   const [searchTerm, setSearchTerm] = useState<string>('');
+  const [currentPage, setCurrentPage] = useState<number>(1);
+  const itemsPerPage = 10;
   const navigate = useNavigate();
 
   useEffect(() => {
     const fetchProducts = async () => {
       try {
-        // Fetch categories and subcategories
         const categorySnapshot = await getDocs(collection(db, 'category'));
         const subCategorySnapshot = await getDocs(collection(db, 'subcategory'));
 
@@ -57,32 +58,20 @@ const ProductList: React.FC = () => {
         setCategories(categoryList);
         setSubCategories(subCategoryList);
 
-        // Fetch products
         const productSnapshot = await getDocs(collection(db, 'product'));
         const productList: Product[] = productSnapshot.docs.map(doc => {
           const data = doc.data();
-          let addedAt = data.added_at;
-
-          // Convert string-based timestamp to a Unix timestamp (if needed)
-          if (typeof addedAt === 'string') {
-            addedAt = new Date(addedAt).getTime();
-          }
-
           return {
             id: doc.id,
             productName: data.productName,
             productCategory: data.productCategory,
             productSpecCategory: data.productSpecCategory,
             productImage: data.productImage,
-            added_at: addedAt,
           };
         });
 
-        // Sort products by added_at in descending order (newest first)
-        productList.sort((a, b) => b.added_at - a.added_at);
-
         setProducts(productList);
-        setFilteredProducts(productList); // Initialize filtered products
+        setFilteredProducts(productList);
       } catch (error) {
         console.error("Error fetching products:", error);
         setError('Échec de la récupération des données');
@@ -100,15 +89,25 @@ const ProductList: React.FC = () => {
         searchTerm === '' || product.productName.toLowerCase().includes(searchTerm.toLowerCase())
       )
     );
+    setCurrentPage(1); 
   }, [searchTerm, products]);
 
+  const indexOfLastProduct = currentPage * itemsPerPage;
+  const indexOfFirstProduct = indexOfLastProduct - itemsPerPage;
+  const currentProducts = filteredProducts.slice(indexOfFirstProduct, indexOfLastProduct);
+
+  const paginate = (pageNumber: number) => setCurrentPage(pageNumber);
+
   const handleDelete = async (productId: string) => {
-    try {
-      await deleteDoc(doc(db, 'product', productId));
-      setProducts(products.filter(product => product.id !== productId));
-      setFilteredProducts(filteredProducts.filter(product => product.id !== productId));
-    } catch (error) {
-      setError('Échec de la suppression du produit');
+    const confirmation = window.confirm("Êtes-vous sûr de vouloir supprimer ce produit ?");
+    if (confirmation) {
+      try {
+        await deleteDoc(doc(db, 'product', productId));
+        setProducts(products.filter(product => product.id !== productId));
+        setFilteredProducts(filteredProducts.filter(product => product.id !== productId));
+      } catch (error) {
+        setError('Échec de la suppression du produit');
+      }
     }
   };
 
@@ -126,6 +125,51 @@ const ProductList: React.FC = () => {
     return subCategory ? subCategory.name : 'Inconnu';
   };
 
+  const renderPagination = () => {
+    const totalPages = Math.ceil(filteredProducts.length / itemsPerPage);
+    const paginationItems = [];
+    if (totalPages <= 1) return null;
+    paginationItems.push(
+      <IconButton
+        key={1}
+        className={`pagination-button ${currentPage === 1 ? 'active' : ''}`}
+        onClick={() => paginate(1)}
+      >
+        1
+      </IconButton>
+    );
+
+    if (currentPage > 3) {
+      paginationItems.push(<span key="start-ellipsis">...</span>);
+    }
+    for (let i = Math.max(2, currentPage - 1); i <= Math.min(totalPages - 1, currentPage + 1); i++) {
+      paginationItems.push(
+        <IconButton
+          key={i}
+          className={`pagination-button ${currentPage === i ? 'active' : ''}`}
+          onClick={() => paginate(i)}
+        >
+          {i}
+        </IconButton>
+      );
+    }
+    if (currentPage < totalPages - 2) {
+      paginationItems.push(<span key="end-ellipsis">...</span>);
+    }
+    if (totalPages > 1) {
+      paginationItems.push(
+        <IconButton
+          key={totalPages}
+          className={`pagination-button ${currentPage === totalPages ? 'active' : ''}`}
+          onClick={() => paginate(totalPages)}
+        >
+          {totalPages}
+        </IconButton>
+      );
+    }
+    return paginationItems;
+  };
+
   if (loading) {
     return <Typography>Chargement...</Typography>;
   }
@@ -134,10 +178,14 @@ const ProductList: React.FC = () => {
     return <Typography color="error">{error}</Typography>;
   }
 
+  const handleViewDetails = (productId: string) => {
+    navigate(`/ReviewsList/${productId}`);
+  };
+  
   return (
     <Container className="product-list-container">
       <Typography variant="h4" gutterBottom>
-        Liste des Produits
+        Liste des produits
       </Typography>
       <div className="search-export-container">
         <form className="form-search" onSubmit={(e) => e.preventDefault()}>
@@ -156,67 +204,99 @@ const ProductList: React.FC = () => {
               <i className="icon-search"></i>
             </button>
           </div>
-        </form>
-        
+        </form>     
         <div className="add-buttons">
-          <Button
-            id='exporter'
-            variant="contained"
-            className="add-button"
-            onClick={() => navigate('/categories')}
-          >
-            Liste Catégories
-          </Button>
-          <Button
-            id='exporter'
-            variant="contained"
-            className="add-button"
-            onClick={() => navigate('/products/new')}
-          >
-            + Produit
-          </Button>
+        <Tooltip className='custom-tooltip' title="Voir la liste des catégories" arrow>
+        <Button
+          id="exporter"
+          variant="contained"
+          className="add-button"
+          onClick={() => navigate('/categories')}
+        > Liste Catégories </Button>
+      </Tooltip>
+      <Tooltip className='custom-tooltip' title="Ajouter un nouveau produit" arrow>
+        <Button
+          id="exporter"
+          variant="contained"
+          className="add-button"
+          onClick={() => navigate('/products/new')}
+        > + Produit </Button>
+      </Tooltip>
         </div>
       </div>
-      <div className="table-container">
-        <table id='ho' className="product-table">
-          <thead>
-            <tr className="table-title">
-              <th>Image du Produit</th>
-              <th>Nom</th>
-              <th>Catégorie</th>
-              <th>Sous-Catégorie</th>
-              <th>Action</th>
-            </tr>
-          </thead>
-          <tbody>
-            {filteredProducts.map((product, index) => (
-              <tr key={product.id} className={`product-item ${index % 2 === 0 ? 'even-row' : 'odd-row'}`}>
-                <td>
-                  <img src={product.productImage} alt={product.productName} className="product-image" />
-                </td>
-                <td>{product.productName}</td>
-                <td>{getCategoryName(product.productCategory)}</td>
-                <td>{getSubCategoryName(product.productSpecCategory)}</td>
-                <td>
-                  <IconButton className="action-button" onClick={() => handleUpdate(product.id)}>
-                    <Edit />
-                  </IconButton>
-                  <IconButton className="action-button delete-button" onClick={() => handleDelete(product.id)}>
-                    <Delete />
-                  </IconButton>
-                </td>
-              </tr>
-            ))}
-          </tbody>
-        </table>
-      </div>
-
+      <TableContainer component={Paper} className="table-container">
+      <Table className="custom-table">
+        <TableHead className="custom-table-head">
+          <TableRow>
+            <TableCell className="custom-table-cell-list" sx={{ fontSize: '14px', fontWeight: '600', textAlign: 'center' }} >Image</TableCell>
+            <TableCell className="custom-table-cell-list" sx={{ fontSize: '14px', fontWeight: '600', textAlign: 'center' }}>Nom</TableCell>
+            <TableCell className="custom-table-cell-list" sx={{ fontSize: '14px', fontWeight: '600', textAlign: 'center' }}>Catégorie</TableCell>
+            <TableCell className="custom-table-cell-list" sx={{ fontSize: '14px', fontWeight: '600', textAlign: 'center' }}>Sous-Catégorie</TableCell>
+            <TableCell className="custom-table-cell-list" sx={{ fontSize: '14px', fontWeight: '600', textAlign: 'center' }}>Action</TableCell>
+          </TableRow>
+        </TableHead>
+        <TableBody>
+          {filteredProducts.map((product, index) => (
+            <TableRow key={product.id} className={`custom-table-row ${index % 2 === 0 ? 'even-row' : 'odd-row'}`}>
+              <TableCell sx={{ textAlign: 'center' }}>
+                <img src={product.productImage} alt={product.productName} className="product-image" />
+              </TableCell>
+              <TableCell className="custom-table-cell-cellule" sx={{
+                fontSize: {
+                  sm:'16px',
+                  md: '16px', 
+                },
+                textAlign: 'center',
+              }}>{product.productName}</TableCell>
+              <TableCell className="custom-table-cell-cellule" sx={{
+                fontSize: { 
+                  sm:'16px',
+                  md: '16px', 
+                },
+                textAlign: 'center',
+              }}>{getCategoryName(product.productCategory)}</TableCell>
+              <TableCell className="custom-table-cell-cellule" sx={{
+                fontSize: {
+                  sm:'16px',
+                  md: '16px', 
+                },
+                textAlign: 'center',
+              }}>{getSubCategoryName(product.productSpecCategory)}</TableCell>
+              <TableCell sx={{ textAlign: 'center' }}>
+                <IconButton className="action-button" onClick={() => handleUpdate(product.id)}>
+                  <Edit />
+                </IconButton>
+                <IconButton className="action-button delete-button" onClick={() => handleDelete(product.id)}>
+                  <Delete />
+                </IconButton>
+                <div 
+                  className="view-details-icon" 
+                  onClick={() => handleViewDetails(product.id)} 
+                  style={{ color:'#8c8f8f', textAlign: 'center' , cursor:'pointer' , display:'inline-block'}} >
+                    <FaEye />
+                </div>
+              </TableCell>
+            </TableRow>
+          ))}
+        </TableBody>
+      </Table>
+    </TableContainer>
       <div className="pagination-container">
-        <IconButton className="pagination-button">&lt;</IconButton>
-        <IconButton className="pagination-button active">1</IconButton>
-        <IconButton className="pagination-button">2</IconButton>
-        <IconButton className="pagination-button">3</IconButton>
-        <IconButton className="pagination-button">&gt;</IconButton>
+        <IconButton
+          className="pagination-button"
+          onClick={() => paginate(currentPage - 1)}
+          disabled={currentPage === 1}
+        >
+          &lt;
+        </IconButton>
+        {renderPagination()}
+        <IconButton
+          className="pagination-button"
+          onClick={() => paginate(currentPage + 1)}
+          disabled={currentPage === Math.ceil(filteredProducts.length / itemsPerPage)}
+        >
+          &gt;
+        </IconButton>
       </div>
     </Container>
   );
